Changes to sparse checkout
The existing code checkout downloaded more data than it needed to.
This is an attempt to make it more sparse and to fix some potential
bugs.

Turn `CHECKOUT_PATHS` into an array so that multiple paths can be used.

Add a step to scan SSH keys of the remote git host.

Add `--filter=blob:none` to the `git clone` so that file contents are
not downloaded until needed by git.

If `$BUILDKITE_REPO_MIRROR` is set, then pass it as `--reference` to
`git clone`. This is for users using the Buildkite repository mirror
functionality.

Add `--depth 1` to the `git fetch` commands for a shallow clone to
checkout only necessary data.

Remove the deprecated `git sparse-checkout init` command.

Use `git set` in place of `git add` and pass the `--no-cone` option so
that the paths passed in can match files and patterns.

The `git checkout` needs to happen after the `git sparse-checkout`
command has been run.
Doing it beforehand checkouts out all files in the repository, which
defeats the purpose of performing a sparse checkout.

This is a `bash` script so use `[[ ]]` in place of `[ ]` and there is no
need to quote the variables within `[[ ]]` like there was before.

Quote other uses of the variables to fix warnings from shellcheck.